requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.9.2
gTTS==2.5.3
faster-whisper==1.0.3
numpy==1.26.4
soundfile==0.12.1
aiofiles==24.1.0
ffmpeg-python==0.2.0
pydub==0.25.1
PyYAML==6.0.2
setuptools<81
